requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
Pillow==11.2.1
zeroconf==0.147.2
